Move functions and update data.sql
